requirements.txt
# Core dependencies
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
faiss-cpu>=1.7.4
numpy>=1.24.0
openai>=1.0.0
groq>=0.4.0
tiktoken>=0.5.0
langchain-text-splitters>=0.0.1
pydantic>=2.0.0
python-dotenv>=1.0.0

# API dependencies
fastapi>=0.104.0
uvicorn>=0.24.0

